fix: data access when having non-zero metadata size
The function get_data_ptr_from_fragment() actually returns a pointer
to the metadata header. In most of the code it is the expected behavior but in
fragments_to_string() we want the actual content without the metadata.
Creation of a new function prefixed by _ to address this particular case.